requirements.txt
scikit-surprise>=1.1.1
pandas>=1.5.3
numpy>=1.23.0
matplotlib>=3.6.3
seaborn>=0.11.2
scipy>=1.10.0
scikit-learn>=1.2.0
